Fix issue #2
The action to get the clipboard was not given enough time.
If this solution doesn't fix this problem for everyone, another would be to keep a copy of the previous clipboard, compare it, and retry with a delay a couple of times to ensure it's not the old clipboard by mistake.