DOES NOT BUILD!! WORK IN PROGRESS
At the moment the Sidekick window forms do not communicate with the new projects at all.

Added an editorconfig file to enforce the tab length to 4.
Started refactoring and restructuring the code.
Created the Sidekick.Core project to facilitate the configuration of the DependencyInjection mecanism.
We can use attributes to register our services with the provider.
Created the Sidekick.Business project.
Moved the ItemParser, Items and Language classes.
Isolated each class into its own file when moved to Sidekick.Business.