#97
First working build with the new structure.
There are a ton of shortcuts taken to have a build working quickly.
Some classes previously moved into the Sidekick.Business have been moved into slightly different folders.
Moved some Setting enums into the Core project so that they can be used everywhere.
Had to tweak the notifications so that the Business could send Hotkeys without having the values of the hotkeys. The hotkeys are currently only known in the UI project. So we pass a string like "Press {{KeybindSetting:CloseWindow}} instead of the value. Inside the UI project we then translate that into the real keybind.