#157 Added the new localization to the settings view.
We are now using Resource (resx) files. There are lots of advantages as explained in the original issue.
The drawback is that before drawing the view,we need to set the Culture every time. This is done through the base SidekickViewModel class.

The settings are now storing a two letter value of the culture (en, fr, de). This makes it easier for us to set the Thread Culture.

The old implementation of IUILanguage now fallsback to English. The overlay view was causing issues because of all the throw exceptions.