Fixed Ctrl+F keybind
Keybind events now need to return a boolean value to let the NativeService know if it has been handled or not. For some specific keybinds, we can wait on the result of that task and resend the input in the case of specific keybinds.

The KeybindEvents has been cleaned up so it is easier to read and maintain